# docs/swagger.yaml
basePath: /api/v2
definitions:
...
  factrak.AgreementCreateParams:
    properties:
      agree:
        type: boolean
    type: object
  factrak.AgreementUpdateParams:
    properties:
      agree:
        type: boolean
    type: object
...
  models.AreaOfStudy:
    properties:
      abbreviation:
        type: string
      courses:
        description: Has many courses
        items:
          $ref: '#/definitions/models.Course'
        type: array
      department:
        $ref: '#/definitions/models.Department'
      departmentID:
        description: Belongs to department
        type: integer
      id:
        type: integer
      name:
        type: string
      professors:
        description: Many2Many professors  (computed periodically from courses)
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.BannedUser:
    properties:
      bulletinRead:
        type: boolean
      bulletinWrite:
        type: boolean
      dormtrak:
        type: boolean
      ephcatch:
        type: boolean
      ephmatch:
        type: boolean
      factrak:
        description: If banned from this service. True means user is banned; false
          means user is not banned
        type: boolean
      id:
        type: integer
      reason:
        description: Reason why user was banned
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Belongs to user
      userID:
        type: integer
    type: object
  models.Book:
    properties:
      authors:
        description: The names of the authors and/or editors for this book.
        items:
          type: string
        type: array
      bookListings:
        description: Has many Book Listings
        items:
          $ref: '#/definitions/models.BookListing'
        type: array
      courses:
        description: Many2Many courses
        items:
          $ref: '#/definitions/models.Course'
        type: array
      id:
        type: integer
      imageLink:
        description: Image link for book cover (Medium)
        type: string
      infoLink:
        description: URL to view information about this book on the Google Books site.
        type: string
      isbn13:
        description: ISBN-13 of this book (all books have this)
        type: string
      publisher:
        description: Publisher of this book.
        type: string
      subtitle:
        description: Book subtitle.
        type: string
      title:
        description: Book title.
        type: string
    type: object
...
  models.Course:
    properties:
      areaOfStudy:
        $ref: '#/definitions/models.AreaOfStudy'
      areaOfStudyID:
        description: Belongs to area of study
        type: integer
      books:
        description: Many2Many books
        items:
          $ref: '#/definitions/models.Book'
        type: array
      factrakScore:
        type: number
      factrakSurveys:
        description: Has many factrak surveys
        items:
          $ref: '#/definitions/models.FactrakSurvey'
        type: array
      id:
        type: integer
      number:
        type: string
      professors:
        description: Has many professors (thru factrak surveys). Ignore this in Gorm.
          Populate this only whenever needed.
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.Department:
    properties:
      areasOfStudy:
        description: Has many areas of study
        items:
          $ref: '#/definitions/models.AreaOfStudy'
        type: array
      id:
        type: integer
      name:
        type: string
      users:
        description: Has many professors & staff
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
...
  models.DormRoom:
    properties:
      area:
        type: integer
      bathroomDesc:
        type: string
      bedAdjustable:
        type: boolean
      closet:
        type: string
      commonRoomAccess:
        type: boolean
      commonRoomDesc:
        type: string
      dorm:
        $ref: '#/definitions/models.Dorm'
      dormID:
        description: Belongs to dorm
        type: integer
      dormtrakReviews:
        description: Has many Dormtrak reviews
        items:
          $ref: '#/definitions/models.DormtrakReview'
        type: array
      faces:
        type: string
      floorNumber:
        type: integer
      flooring:
        type: string
      hc:
        type: boolean
      id:
        type: integer
      keyOrCard:
        type: string
      noise:
        type: string
      number:
        type: string
      picture:
        type: string
      privateBathroom:
        type: boolean
      roomType:
        type: string
      thermostatAccess:
        type: boolean
      users:
        description: Has many students
        items:
          $ref: '#/definitions/models.User'
        type: array
      walkthrough:
        type: boolean
    type: object
...
  models.Ephcatch:
    properties:
      createdTime:
        description: 'Pass the created time: not looked at by GORM'
        type: string
      id:
        type: integer
      other:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Belongs to other
      otherID:
        type: integer
      seen:
        type: boolean
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Belongs to user
      userID:
        type: integer
    type: object
...
  models.FactrakAgreement:
    properties:
      agrees:
        type: boolean
      factrakSurvey:
        $ref: '#/definitions/models.FactrakSurvey'
      factrakSurveyID:
        description: Belongs to survey
        type: integer
      id:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      userID:
        description: Belongs to user
        type: integer
    type: object
  models.FactrakSurvey:
    properties:
      agreements:
        description: Has many agreements
        items:
          $ref: '#/definitions/models.FactrakAgreement'
        type: array
      approachability:
        type: integer
      clientAgreement:
        description: |-
          Pass if the client agreed with the survey; not looked at by GORM.
          True means user agreed, false means user disagreed, and null/missing means user does not have any
          agreement/disagreement.
        type: boolean
      comment:
        type: string
      course:
        $ref: '#/definitions/models.Course'
      courseFormat:
        description: Remote, Hybrid, In-Person
        type: string
      courseID:
        description: Belongs to course
        type: integer
      courseStimulating:
        type: integer
      courseWorkload:
        type: integer
      createdTime:
        description: 'Pass the created time: not looked at by GORM'
        type: string
      flagged:
        type: boolean
      gradeReceived:
        type: string
      id:
        type: integer
      leadLecture:
        type: integer
      mentalHealthSupport:
        type: integer
      outsideHelpfulness:
        type: integer
      professor:
        $ref: '#/definitions/models.User'
      professorID:
        description: Belongs to professor
        type: integer
      promoteDiscussion:
        type: integer
      semesterSeason:
        description: Course info data
        type: string
      semesterYear:
        type: integer
      totalAgree:
        description: Not looked at by GORM, just for returning in JSON
        type: integer
      totalDisagree:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      userID:
        description: Belongs to user (student)
        type: integer
      wouldRecommendCourse:
        type: boolean
      wouldTakeAnother:
        type: boolean
    type: object
...
  models.User:
    properties:
      admin:
        type: boolean
      areasOfStudy:
        description: Many2Many Area of Studies iff professor (computed periodically
          from courses)
        items:
          $ref: '#/definitions/models.AreaOfStudy'
        type: array
      atWilliams:
        type: boolean
      bannedUser:
        allOf:
        - $ref: '#/definitions/models.BannedUser'
        description: Has one (or zero) banned users
      campusPhoneEXT:
        description: campus phone extension after 413-597-
        type: string
      campusStatus:
        description: 'Has campus status: either remote or on-campus'
        type: string
      cellPhone:
        type: string
      classYear:
        type: integer
      department:
        $ref: '#/definitions/models.Department'
      departmentID:
        description: Belongs to Department iff professor (fetched from Williams LDAP)
        type: integer
      dormRoom:
        $ref: '#/definitions/models.DormRoom'
      dormRoomID:
        description: Belongs to Dorm Room iff student
        type: integer
      dormVisible:
        type: boolean
      dormtrakReviews:
        description: Has many dormtrak reviews
        items:
          $ref: '#/definitions/models.DormtrakReview'
        type: array
      entry:
        type: string
      ephcatchEligibility:
        type: boolean
      ephcatches:
        description: Has many ephcatches (owner side)
        items:
          $ref: '#/definitions/models.Ephcatch'
        type: array
      ephmatchProfile:
        allOf:
        - $ref: '#/definitions/models.EphmatchProfile'
        description: Has one ephmatch profile
      factrakAdmin:
        type: boolean
      factrakAgreements:
        description: Has many factrak agreements
        items:
          $ref: '#/definitions/models.FactrakAgreement'
        type: array
      factrakScore:
        description: Only for professor retrieval with rankings
        type: number
      factrakSurveyDeficit:
        type: integer
      factrakSurveys:
        description: |-
          As we cannot be both a student and a professor, this combines either a student or a professor's factrak survey.
          We populate this field as a hook AfterFind.
        items:
          $ref: '#/definitions/models.FactrakSurvey'
        type: array
      hasAcceptedDormtrakPolicy:
        type: boolean
      hasAcceptedFactrakPolicy:
        type: boolean
      homeCountry:
        type: string
      homePhone:
        type: string
      homeState:
        type: string
      homeTown:
        type: string
      homeVisible:
        type: boolean
      homeZip:
        type: string
      id:
        type: integer
      major:
        type: string
      name:
        type: string
      nickname:
        type: string
      notificationSettings:
        allOf:
        - $ref: '#/definitions/models.NotificationSettings'
        description: Has one notification settings
      notificationTokens:
        description: Has many notification tokens
        items:
          $ref: '#/definitions/models.NotificationToken'
        type: array
      offCampus:
        description: If user is Off-Campus and thus doesn't have a dorm
        type: boolean
      offCycle:
        type: boolean
      office:
        $ref: '#/definitions/models.Office'
      officeID:
        description: Belongs to Office iff staff/professor
        type: integer
      onCampusSemester:
        description: used to calculate number of factrack surveys needed
        type: integer
      optOutEphcatch:
        type: boolean
      pronoun:
        type: string
      suBox:
        type: string
      tags:
        description: Has many tags
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      title:
        type: string
      type:
        type: string
      unixID:
        type: string
      visible:
        type: boolean
      williamsEmail:
        type: string
      williamsID:
        description: Williams W# ID
        type: string
    type: object
...
  services.BaseErrorResponse:
    properties:
      error:
        $ref: '#/definitions/services.RespError'
      status:
        type: integer
      updateToken:
        type: boolean
    type: object
  services.RespError:
    properties:
      errorCode:
        type: integer
      errors:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
...
host: wso.williams.edu
info:
  contact:
    email: wso-dev@wso.williams.edu
    name: WSO Dev
  description: API for WSO services like factrak, facebook, dormtrak, course scheduler,
    and others.
  title: WSO API
  version: 2.0.0
paths:
...
  /factrak/areas-of-study:
    get:
      consumes:
      - application/json
      description: lists all areas of study
      operationId: listAreasOfStudy
      parameters:
      - description: Offset Pagination
        in: query
        name: offset
        type: integer
      - description: Limit Pagination
        in: query
        name: limit
        type: integer
      - collectionFormat: csv
        description: Preload List (department, courses)
        in: query
        items:
          type: string
        name: preload
        type: array
      - description: Sort Order (id, name; default name)
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AreaOfStudy'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
      security:
      - Bearer: []
      summary: List areas of study
      tags:
      - factrak
  /factrak/areas-of-study/{areaOfStudyID}:
    get:
      consumes:
      - application/json
      description: get one area of study with department preload
      operationId: getAreaOfStudy
      parameters:
      - description: Area of Study ID
        in: path
        name: areaOfStudyID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AreaOfStudy'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
      security:
      - Bearer: []
      summary: Get area of study
      tags:
      - factrak
  /factrak/areas-of-study/{areaOfStudyID}/courses:
    get:
      consumes:
      - application/json
      deprecated: true
      description: |-
        list one area of study's courses and the course's professors.
        Please note: unlike other GetCourses endpoints, this one also returns a list of each courses professors.
      operationId: listAreaOfStudyCourses
      parameters:
      - description: Area of Study ID
        in: path
        name: areaOfStudyID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Course'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
      security:
      - Bearer: []
      summary: List area of study courses
      tags:
      - factrak
  /factrak/areas-of-study/{areaOfStudyID}/professors:
    get:
      consumes:
      - application/json
      deprecated: true
      description: list one area of study's professors
      operationId: listAreaOfStudyProfessors
      parameters:
      - description: Area of Study ID
        in: path
        name: areaOfStudyID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
      security:
      - Bearer: []
      summary: List area of study professors
      tags:
      - factrak
...
  /factrak/surveys/{surveyID}/agreement:
    delete:
      consumes:
      - application/json
      description: deletes an agreement about a survey made by the self user
      operationId: deleteSurveyAgreement
      parameters:
      - description: Survey ID
        in: path
        name: surveyID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FactrakAgreement'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
        "1551":
          description: survey agreement could not be found
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
      security:
      - Bearer: []
      summary: Delete survey agreement
      tags:
      - factrak
    get:
      consumes:
      - application/json
      description: gets the agreement about a survey made by the self user
      operationId: getSurveyAgreement
      parameters:
      - description: Survey ID
        in: path
        name: surveyID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FactrakAgreement'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
        "1551":
          description: survey agreement could not be found
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
      security:
      - Bearer: []
      summary: Get survey agreement by self
      tags:
      - factrak
    patch:
      consumes:
      - application/json
      description: updates an agreement about a survey made by the self user
      operationId: updateSurveyAgreement
      parameters:
      - description: Survey ID
        in: path
        name: surveyID
        required: true
        type: integer
      - description: Update Agreement Params
        in: body
        name: updateParams
        required: true
        schema:
          $ref: '#/definitions/factrak.AgreementUpdateParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FactrakAgreement'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
        "1100":
          description: could not parse malformed request data
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
        "1551":
          description: survey agreement could not be found
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
      security:
      - Bearer: []
      summary: Update survey agreement
      tags:
      - factrak
    post:
      consumes:
      - application/json
      description: creates an agreement about a survey made by the self user
      operationId: createSurveyAgreement
      parameters:
      - description: Survey ID
        in: path
        name: surveyID
        required: true
        type: integer
      - description: Create Agreement Params
        in: body
        name: createParams
        required: true
        schema:
          $ref: '#/definitions/factrak.AgreementCreateParams'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.FactrakAgreement'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
        "1100":
          description: could not parse malformed request data
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
        "1552":
          description: survey agreement already exists for this user and survey
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
        "1553":
          description: cannot create survey agreement with your own survey
          schema:
            $ref: '#/definitions/services.BaseErrorResponse'
      security:
      - Bearer: []
      summary: Create survey agreement
      tags:
      - factrak
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"